requirements.txt
appdirs==1.4.4
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
astroid==3.0.1
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2023.10
cryptography==41.0.5
cssselect2==0.7.0
dill==0.3.7
dj-database-url==2.1.0
Django==4.2
django-crispy-forms==2.1
django-simple-history==3.4.0
djangorestframework==3.14.0
gunicorn==21.2.0
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
isort==5.12.0
lxml==4.9.3
mccabe==0.7.0
oscrypto==1.3.0
packaging==23.2
pdfkit==1.0.0
Pillow==10.1.0
platformdirs==3.11.0
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg2-binary==2.9.9
pycparser==2.21
pyee==8.2.2
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
pylint==3.0.2
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pypdf==3.17.0
pypng==0.20220715.0
pyppeteer==1.0.2
python-bidi==0.4.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
tomlkit==0.12.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritools==4.0.2
urllib3==1.26.18
webencodings==0.5.1
websockets==10.4
whitenoise==6.6.0
xhtml2pdf==0.2.11
zipp==3.17.0
